Improve config handling
Fix a couple bugs. Having the default values in code was cumbersome, so
use the example file for default values. The example was also not
getting updated, so hopefully using that as the default will keep it
more up to date. Remove some of the plugins that aren't well supported
from the example, and add missing values.